Moving responsibilities from WindshaftMap to Map (WIP)
