+++ src/geo/map/cartodb-layer.js
@@ -98,8 +98,13 @@ var CartoDBLayer = MapLayer.extend({
 
   // Layers inside a "layergroup" layer have the layer_name defined in options.layer_name
   // Layers inside a "namedmap" layer have the layer_name defined in the root of their definition
-  getName: function() {
+  getName: function () {
     return this.get('options') && this.get('options').layer_name || this.get('layer_name');
+  },
+
+  needsRefresh: function () {
+    // TODO: We will need check if the filter for this layer (dataprovider) has changed here
+    return this.hasChanged('visible') || this.hasChanged('sql') || this.hasChanged('cartocss');
   }
 });
