+++ src/windshaft/windshaft-map.js
@@ -1,9 +1,10 @@
+var Backbone = require('backbone');
 var _ = require('underscore');
 var WindshaftFiltersCollection = require('./filters/collection');
-var WindshaftFiltersBoundingBoxFilter = require('./filters/bounding-box');
-var WindshaftMapInstance = require('./windshaft-map-instance');
 var WindshaftLayerGroupConfig = require('./layergroup-config');
 var WindshaftNamedMapConfig = require('./namedmap-config');
+var WindshaftConfig = require('./config');
+var EMPTY_GIF = 'data:image/gif;base64,R0lGODlhAQABAIAAAAAAAP///yH5BAEAAAAALAAAAAABAAEAAAIBRAA7';
 
 /**
  * This class represents the concept of a map in Windshaft. It holds a reference
@@ -13,134 +14,267 @@ var WindshaftNamedMapConfig = require('./namedmap-config');
  * NOTE: We couldn't use a Backbone.Model for this because this class uses a client
  * which is a bit smart about how requests to Windshaft shoud be made.
  */
-var WindshaftMap = function (options) {
-  var BOUNDING_BOX_FILTER_WAIT = 500;
+var WindshaftMap = Backbone.Model.extend({
 
-  this._type = options.type;
-  this.layers = options.layers;
-  this.dataviews = options.dataviews;
-  this.map = options.map;
-  this.client = options.client;
-  this.statTag = options.statTag;
-  this.configGenerator = options.configGenerator;
-  this.instance = new WindshaftMapInstance();
+  initialize: function (attrs, options) {
+    this.client = options.client;
+    this.statTag = options.statTag;
+    this.configGenerator = options.configGenerator;
 
-  this.map.bind('change:center change:zoom', _.debounce(this._boundingBoxChanged, BOUNDING_BOX_FILTER_WAIT), this);
-  this.layers.bind('change', this._layerChanged, this);
-  this.dataviews.bind('change:filter', this._dataviewChanged, this);
-  this.dataviews.bind('add', _.debounce(this._dataviewChanged, 10), this);
+    // TODO: What params are really used?
+    this.pngParams = ['map_key', 'api_key', 'cache_policy', 'updated_at'];
+    this.gridParams = ['map_key', 'api_key', 'cache_policy', 'updated_at'];
+  },
 
-  // delay initial instance creation until sure there are not any dataviews added that will re-create the instance
-  // TODO tmp fix; this is far from ideal and prone to timing issues, can we do something more elegant?
-  var timeout = setTimeout(function () {
-    this._createInstance();
-  }.bind(this), 50);
-  this.dataviews.once('all', function () {
-    // Dataviews incoming, skip initial instance creation
-    clearTimeout(timeout);
-  }, this);
-};
+  isNamedMap: function () {
+    return this.configGenerator === WindshaftNamedMapConfig;
+  },
 
-WindshaftMap.prototype.isNamedMap = function () {
-  return this.configGenerator === WindshaftNamedMapConfig;
-};
+  isAnonymousMap: function () {
+    return this.configGenerator === WindshaftLayerGroupConfig;
+  },
 
-WindshaftMap.prototype.isAnonymousMap = function () {
-  return this.configGenerator === WindshaftLayerGroupConfig;
-};
+  createInstance: function (options) {
+    options = options || {};
+    // WindshaftMap knows what types of layers should be sent to Windshaft:
+    var layers = _.select(options.layers, function (layer) {
+      return layer.get('type') === 'CartoDB' || layer.get('type') === 'torque';
+    });
+    var dataviews = options.dataviews;
+    var sourceLayerId = options.sourceLayerId;
 
-WindshaftMap.prototype._createInstance = function (options) {
-  options = options || {};
+    var mapConfig = this.configGenerator.generate({
+      layers: layers,
+      dataviews: dataviews
+    });
 
-  var cfg = this.configGenerator.generate({
-    layers: this.layers.models,
-    dataviews: this.dataviews
-  });
+    var filtersFromVisibleLayers = dataviews.chain()
+      .filter(function (dataview) { return dataview.layer.isVisible(); })
+      .map(function (dataview) { return dataview.filter; })
+      .compact() // not all dataviews have filters
+      .value();
 
-  var filtersFromVisibleLayers = this.dataviews.chain()
-    .filter(function (dataview) { return dataview.layer.isVisible(); })
-    .map(function (dataview) { return dataview.filter; })
-    .compact() // not all dataviews have filters
-    .value();
+    var filters = new WindshaftFiltersCollection(filtersFromVisibleLayers);
 
-  var filters = new WindshaftFiltersCollection(filtersFromVisibleLayers);
+    this.client.instantiateMap({
+      mapDefinition: mapConfig,
+      statTag: this.statTag,
+      filters: filters.toJSON(),
+      success: function (mapInstance) {
+        this.set(mapInstance);
+        this.trigger('instanceCreated', this, sourceLayerId);
 
-  this.client.instantiateMap({
-    mapDefinition: cfg,
-    statTag: this.statTag,
-    filters: filters.toJSON(),
-    success: function (mapInstance) {
-      // Update the map instance with the attributes of the newly created one
-      this.instance.set(mapInstance.toJSON());
+        // TODO: Revisit this (will layerIndex work for NamedMaps??)
+        // Should we move it somewhere else?
+        _.each(layers, function (layer, layerIndex) {
+          if (layer.get('type') === 'torque') {
+            layer.set('meta', mapInstance.getLayerMeta(layerIndex));
+            layer.set('urls', mapInstance.getTiles('torque'));
+          }
+        });
+      }.bind(this),
+      error: function (error) {
+        console.log('Error creating the map instance on Windshaft: ' + error);
+      }
+    });
 
-      // update other kind of layers too
-      this.layers.each(function (layer, layerIndex) {
-        if (layer.get('type') === 'torque') {
-          layer.set('meta', mapInstance.getLayerMeta(layerIndex));
-          layer.set('urls', mapInstance.getTiles('torque'));
+    return this;
+  },
+
+  TILE_EXTENSIONS_BY_LAYER_TYPE: {
+    'mapnik': '.png',
+    'torque': '.json.torque'
+  },
+
+  getBaseURL: function (subhost) {
+    return [
+      this._getHost(subhost),
+      WindshaftConfig.MAPS_API_BASE_URL,
+      this.get('layergroupid')
+    ].join('/');
+  },
+
+  _getHost: function (subhost) {
+    var userName = this.get('userName');
+    var protocol = this._useHTTPS() ? 'https' : 'http';
+    subhost = subhost ? subhost + '.' : '';
+    var host = this.get('urlTemplate').replace('{user}', userName);
+    var cdnHost = this.get('cdn_url') && this.get('cdn_url')[protocol];
+    if (cdnHost) {
+      host = [protocol, '://', subhost, cdnHost, '/', userName].join('');
+    }
+
+    return host;
+  },
+
+  _useHTTPS: function () {
+    return this.get('urlTemplate').indexOf('https') === 0;
+  },
+
+  getDataviewURL: function (options) {
+    var dataviewId = options.dataviewId;
+    var protocol = options.protocol;
+    var url;
+    var layers = this.get('metadata') && this.get('metadata').layers;
+
+    _.each(layers, function (layer) {
+      // TODO layer.widgets is the raw data returned from metadata… should be renamed once the result from Windshaft is changed
+      var dataviews = layer.widgets;
+      for (var id in dataviews) {
+        if (dataviewId === id) {
+          url = dataviews[id].url[protocol];
+          return;
         }
-      });
+      }
+    });
 
-      this._updateDataviewURLs({
-        layerId: options.layerId
-      });
-    }.bind(this),
-    error: function (error) {
-      console.log('Error creating the map instance on Windshaft: ' + error);
+    return url;
+  },
+
+  getTiles: function (layerType, params) {
+    var grids = [];
+    var tiles = [];
+
+    var pngParams = this._encodeParams(params, this.pngParams);
+    var gridParams = this._encodeParams(params, this.gridParams);
+    var subdomains = ['0', '1', '2', '3'];
+
+    if (this._useHTTPS()) {
+      subdomains = [''];
     }
-  });
 
-  return this.instance;
-};
+    layerType = layerType || 'mapnik';
 
-WindshaftMap.prototype._boundingBoxChanged = function () {
-  if (this.instance.isLoaded()) {
-    this._updateDataviewURLs();
+    var layerIndexes = this._getLayerIndexesByType(layerType);
+    if (layerIndexes.length) {
+      var gridTemplate = '/{z}/{x}/{y}';
+
+      for (var i = 0; i < subdomains.length; ++i) {
+        var subdomain = subdomains[i];
+        var tileURLTemplate = [
+          this.getBaseURL(subdomain),
+          '/',
+          layerIndexes.join(','),
+          '/{z}/{x}/{y}',
+          this.TILE_EXTENSIONS_BY_LAYER_TYPE[layerType],
+          (pngParams ? '?' + pngParams : '')
+        ].join('');
+
+        tiles.push(tileURLTemplate);
+
+        // for mapnik layers add grid json too
+        if (layerType === 'mapnik') {
+          for (var layer = 0; layer < this.get('metadata').layers.length; ++layer) {
+            var index = this._getLayerIndexByType(layer, 'mapnik');
+            if (index >= 0) {
+              var gridURLTemplate = [
+                this.getBaseURL(subdomain),
+                '/',
+                index,
+                gridTemplate,
+                '.grid.json',
+                (gridParams ? '?' + gridParams : '')
+              ].join('');
+              grids[layer] = grids[layer] || [];
+              grids[layer].push(gridURLTemplate);
+            }
+          }
+        }
+      }
+    } else {
+      // TODO: Clients of this method should decide what to render if no layers are present
+      tiles = [EMPTY_GIF];
+    }
+
+    this.urls = {
+      tiles: tiles,
+      grids: grids
+    };
+    return this.urls;
+  },
+
+  getLayerMeta: function (layerIndex) {
+    var layerMeta = {};
+    var layers = this.get('metadata') && this.get('metadata').layers;
+    if (layers && layers[layerIndex]) {
+      layerMeta = layers[layerIndex].meta || {};
+    }
+    return layerMeta;
+  },
+
+  _encodeParams: function (params, included) {
+    if (!params) return '';
+    var url_params = [];
+    included = included || _.keys(params);
+    for (var i in included) {
+      var k = included[i];
+      var p = params[k];
+      if (p) {
+        if (_.isArray(p)) {
+          for (var j = 0, len = p.length; j < len; j++) {
+            url_params.push(k + '[]=' + encodeURIComponent(p[j]));
+          }
+        } else {
+          var q = encodeURIComponent(p);
+          q = q.replace(/%7Bx%7D/g, '{x}').replace(/%7By%7D/g, '{y}').replace(/%7Bz%7D/g, '{z}');
+          url_params.push(k + '=' + q);
+        }
+      }
+    }
+    return url_params.join('&');
+  },
+
+  /**
+   * Returns the index of a layer of a given type, as the tiler kwows it.
+   *
+   * @param {string|array} types - Type or types of layers
+   */
+  _getLayerIndexesByType: function (types) {
+    var layers = this.get('metadata') && this.get('metadata').layers;
+
+    if (!layers) {
+      return;
+    }
+    var layerIndexes = [];
+    for (var i = 0; i < layers.length; i++) {
+      var layer = layers[i];
+      var isValidType = false;
+      if (types && types.length > 0) {
+        isValidType = types.indexOf(layer.type) !== -1;
+      }
+      if (isValidType) {
+        layerIndexes.push(i);
+      }
+    }
+    return layerIndexes;
+  },
+
+  /**
+   * Returns the index of a layer of a given type, as the tiler kwows it.
+   *
+   * @param {integer} index - number of layer of the specified type
+   * @param {string} layerType - type of the layers
+   */
+  _getLayerIndexByType: function (index, layerType) {
+    var layers = this.get('metadata') && this.get('metadata').layers;
+
+    if (!layers) {
+      return index;
+    }
+
+    var tilerLayerIndex = {};
+    var j = 0;
+    for (var i = 0; i < layers.length; i++) {
+      if (layers[i].type === layerType) {
+        tilerLayerIndex[j] = i;
+        j++;
+      }
+    }
+    if (tilerLayerIndex[index] === undefined) {
+      return -1;
+    }
+    return tilerLayerIndex[index];
   }
-};
-
-WindshaftMap.prototype._updateDataviewURLs = function (options) {
-  options = options || {};
-  var boundingBoxFilter = new WindshaftFiltersBoundingBoxFilter(this.map.getViewBounds());
-  var boundingBox = boundingBoxFilter.toString();
-  var layerId = options.layerId;
-
-  this.dataviews.each(function (dataview) {
-    var url = this.instance.getDataviewURL({
-      dataviewId: dataview.get('id'),
-      protocol: 'http'
-    });
-
-    var layerMeta = dataview.layer.get('meta') || {};
-    var extraAttrs = {};
-    if (layerMeta.steps && layerMeta.column_type && _.isNumber(layerMeta.start) && _.isNumber(layerMeta.end)) {
-      extraAttrs = {
-        bins: layerMeta.steps,
-        columnType: layerMeta.column_type,
-        start: layerMeta.start / 1000,
-        end: layerMeta.end / 1000
-      };
-    }
-
-    dataview.set(_.extend({
-      'url': url,
-      'boundingBox': boundingBox
-    }, extraAttrs), {
-      silent: layerId && layerId !== dataview.layer.get('id')
-    });
-  }, this);
-};
-
-WindshaftMap.prototype._dataviewChanged = function (dataview) {
-  this._createInstance({
-    layerId: dataview.layer.get('id')
-  });
-};
-
-WindshaftMap.prototype._layerChanged = function (layer) {
-  this._createInstance({
-    layerId: layer.get('id')
-  });
-};
+});
 
 module.exports = WindshaftMap;
